Started testing
If it's in "testing mode", the game plays itself generating random moves
(and taking forever to reach the goal).

Already found errors that are going to be fixed some other day, now that
I found out that this is not due in one hour. (See? Going to the Jay's
office really helps guys!)